+++ Parsedown-5.3.php
@@ -1,0 +1,5 @@
+<?php namespace Parsedown;
+
+require_once 'Parsedown-core.php';
+
+class Parsedown extends \ParsedownCore {}
